% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{bar,Foo-method}
\alias{bar,Foo-method}
\title{A dummy method}
\usage{
\S4method{bar}{Foo}(x)
}
\arguments{
\item{x}{Foo}
}
\description{
A dummy method
}
\seealso{
Other Foo functions: 
\code{\link{Foo-class}}
}
\concept{Foo functions}
