% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{class}
\name{Foo-class}
\alias{Foo-class}
\title{A dummy class}
\description{
A dummy class
}
\section{Slots}{

\describe{
\item{\code{foo}}{character}
}}

\seealso{
Other Foo functions: 
\code{\link{bar,Foo-method}}
}
\concept{Foo functions}
